+++ src/common/entity/sort_entities_by_attribute.py
@@ -3,7 +3,7 @@ from typing import List
 from common.exceptions import BadRequestException
 
 
-def get_value_from_attribute_spec(document: dict, attribute_path_elements: List[str]):
+def _get_value_from_attribute_spec(document: dict, attribute_path_elements: List[str]):
     for attr_key in attribute_path_elements:
         try:
             # Convert to int if data is a list, and attr_key is numeric
@@ -23,5 +23,5 @@ def get_value_from_attribute_spec(document: dict, attribute_path_elements: List[
 
 def sort_dtos_by_attribute(entity_list: List[dict], dotted_attribute_path: str) -> List[dict]:
     attribute_path_elements = dotted_attribute_path.split(".")
-    entity_list.sort(key=lambda document: get_value_from_attribute_spec(document, attribute_path_elements))
+    entity_list.sort(key=lambda document: _get_value_from_attribute_spec(document, attribute_path_elements))
     return entity_list
